+++ CHANGES.md
@@ -12,6 +12,7 @@ Improvements:
 - fix(Pony): improve regex for numbers
 - JSON: support for comments in JSON (#2016)
 - fix(cpp): improve string literal matching
+- fix(Go): improve function matching
 
 [Carl Baxter]: https://github.com/cdbax